Fixes Cmap/Jaccard query blocked
The Cmap and Jaccard REST queries are blocked because the requests wait
for the complete analysis to finish. The cause of this bug is that the
computation methods in the *ComputationService EJBs, that are
responsible for launching the pipeline execution, are synchronous.

This commit fixes this bug by setting these methods as asynchronous by
addint the @Asynchronous annotation.