Fixes Jaccard query to properly use the onlyUniverseGenes parameter
So far, the onlyUniverseGenes was being only used to validate the sizes of the input
gene sets. However, it should be also used when retrieving the gene sets for the
comparisons (both from DB signatures and from the user input) and, therefore, it
should be stored in the JaccardResult entity. This commit fixes this analysis to use
it properly.