Fixes the FDR correction step in the Jaccard query analysis
This commit fixes the FDR corrections step in the Jaccard query analysis so that it
takes into account the possibility that there may be no signatures to compare
against the user input. In this case, the analysis should end properly and without
results for the user input. So far, it was failing and this commit fixes this
wrong failure.