Denormalizes DrugSignatureInteraction table
This commit denormalizes the DrugSignatureInteraction table in order to increase the
performance of de queries with real data (up to > 30 million rows). The denormalized
table is named FullDrugSignatureInteraction and this is the entity used now by the
corresponding Dao, which reconstructs the original entities after performing the
queries.

This entity also declares indexes on several fields in order to improve the
performance of the queries. To make this indexes work, queries using those fields in
where clauses are now compared with "LIKE 'text%'" instead of "LIKE '%text%'". Fields
without indexes are:
	- signatureCellSubTypeA, signatureCellSubTypeB, and signatureDisease:
because these fields may contain multiple values separated by "###".
	- drugSourceDb: because this field is not shown in the frontend.