Fixes a bug when listing the cell type values
This commit fixes a bug ocurred when listing the cell type values which caused "###"
from the full_drug_signature_interaction table to appear in the results.

Also, this commit updates the sample database to include a case where one signature
has multiple cell subtypes ant the correct behaviour of this listing can be checked.