Allow Cmap/Jaccard queries using down genesets
This commit introduces the necessary modifications to allow Cmap/Jaccard queries
using only down genesets. So far, queries were allowed using both up and down
genesets or only up genesets. This way, geneset queries are now identified as "UP"
or "DOWN" depending on the geneset provided for the analyis, a value which is stored
in the corresponding database tables.

Using this information, the CmapDrugGeneSetSignatureInteraction entity is now able
to return the drug effect, a calculated field based on the geneset type (UP/DOWN).