Allow to specify case and reference types in Cmap queries
This commit introduces the necessary to specify case and reference types in Cmap
queries.

To support this feature, these fields were added to the CmapResult entity and are
now accepted as parameters in the REST method by adding them to the
CmapQueryParameters.

Then, these two fields are used to generate a summary for the Cmap results using
the PredictionSummaryGenerator, which has been refactored in this commit. Also,
the Cmap result lists can include these predictions summaries if the
“includeSummary” is passed as query parameter.