Add cell type or subtype filters
This commit introduces support for a new filter on both the type or subtype at the
same time. This new filter is incompatible with the separate cell type and cell
subtype filters. When this filter is used, cells whose type or subtype matches the
filter are allowed to appear in the results.