Add methods to retrieve information associated with the Jaccard results
This commit adds new methods and resources associated with Jaccard results to allow
the creation of new representations in the frontend. A resource to obtain the cell
type and cell subtype distributions among the signatures involved in a Jaccard
analysis has been added. On the other hand, a new resource and service to perform
Hypergeometric tests has also been added.